+++ strawberryfields/backends/gaussianbackend/backend.py
@@ -248,16 +248,15 @@ class GaussianBackend(BaseGaussian):
 
         mu = self.circuit.mean
         cov = self.circuit.scovmatxp()
+        mean = self.circuit.smeanxp()
         # check we are sampling from a gaussian state with zero mean
-        if not allclose(mu, zeros_like(mu)):
-            raise NotImplementedError(
-                "Threshold measurement is only supported for " "Gaussian states with zero mean"
-            )
+
         x_idxs = array(modes)
         p_idxs = x_idxs + len(mu)
         modes_idxs = concatenate([x_idxs, p_idxs])
         reduced_cov = cov[ix_(modes_idxs, modes_idxs)]
-        samples = torontonian_sample_state(reduced_cov, shots)
+        reduced_mean = mean[modes_idxs]
+        samples = torontonian_sample_state(mu=reduced_mean, cov=reduced_cov, samples=shots)
 
         return samples